LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_touch-adjustment_click_target-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Touch letter 'O' below to run the test. If a "PASS" result appears the test passes, otherwise it fails

Link O


Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT Touch-generated events should have the same target Test timed out

Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT Touch-generated events should have the same target Test timed out
